Move abilities for admin views in separate model
